requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-cors-headers==2.5.3
django-extensions==2.1.6
djangorestframework==3.9.3
entrypoints==0.3
flake8==3.7.7
idna==2.8
mccabe==0.6.1
mypy==0.701
mypy-extensions==0.4.1
Pillow==6.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.25.3
